+++ resources/ReleaseNotes.md
@@ -1,1 +1,3 @@
-- Fix issue with misalignment applied to wrong axis
+- Convert data ready time to pseudo-timestamp
+- Correct FOB and FIB temperature conversion, based on latest sensor calibration
+- Fix issue in calibration data load when sensor model is empty